root-overlayfs: mount overlay on top of /sysroot
This module performs a couple of operations at boot:
1. creates a new partition with the remaining disk space. The partition
   will be LUKS encrypted.
2. mount the partition in /run/data
3. mount an overlayfs on /sysroot where upper and work are in the
   encrypted partition.

This is especially useful for systems that use dm-verity protetion,
meaning the whole root disk is read-only and writes are not allowed.
By creating a separate partition on the fly at boot, we ensure that the
encryption partition cannot contain injected files and at the same time
keep the root pristine. Thanks to the overlay, we don't need to modify
any process running in the OS, because the root will just look as a
normal rw disk.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>