systemd-copy-initrd-modules: copy modules from /lib/modules into /run/modules
This module simply copies all content of $initd/lib/modules
into /run/modules. Since /run is not destroyed during the switch_root,
it can be then used by libkmod to find kernel modules *if* /lib/modules
of the rootfs is empty.

This module is especially useful in environments that boot with UKIs
(Unified kernel image).

The advantage of this approach is that we can reuse the same set of
minimal modules from initrd to the rootfs, without the need of
installing additional modules (but leaving the user the possibility to
do so). In environments like UKIs where the initramfs is tied to the
kernel and signed & shipped together, this guarantees a minimal set of
modules to boot and use the system, reducing the amount of modules to trust to
the minimum. Having an UKI also eliminates the possibility of
copying mismatching initramfs modules into rootfs, as each initramfs is
built and tied to the same kernel in the UKI, and therefore updating a
kernel means also updating the UKI.

Of course if the user wants to install additional modules, the ones
copied in /run by this module will be automatically discarded.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>